+++ tests/Stubs/Migrations/2020_01_23_234567_create_table_dogs.php
@@ -1,0 +1,35 @@
+<?php
+
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+/**
+ * Class CreateTableDogs
+ */
+class CreateTableDogs extends \Illuminate\Database\Migrations\Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('dogs', function (Blueprint $table) {
+            $table->bigIncrements('id');
+            $table->string('name');
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('dogs');
+    }
+}