+++ tests/Stubs/Repositories/PersonRepository.php
@@ -1,0 +1,16 @@
+<?php
+
+
+namespace bigpaulie\repository\tests\Stubs\Repositories;
+
+
+use bigpaulie\repository\AbstractRepository;
+
+/**
+ * Class PersonRepository
+ * @package bigpaulie\repository\tests\Stubs
+ */
+class PersonRepository extends AbstractRepository
+{
+
+}