+++ changelog.md
@@ -2,6 +2,12 @@
 
 All notable changes to `Repository` will be documented in this file.
 
+## Version 2.0.0
+- Refactoring AbstractRepository class
+- Refactoring tests
+- Adding helper function `repository()`
+- Adding repository generator command
+
 ## Version 1.0
 
 ### Added